requirements.txt
flask
torch
torchvision
pillow
werkzeug
flask-pymongo
opencv-python
flask-cors
